t03_client_server/queue_manager/requirements.txt
# python2!
../../mns_python_sdk
pyzmq==17.0.0
